- 0.5 little fixes

+++ branches/working-0.5/src/request_parse.c
@@ -77,7 +77,7 @@ static char *head_to_upper(char *str)
     char *st = str;
     while (*str) {
         if (*str=='-') *str = '_';
-        else *str = (char)toupper(*str);
+        else *str = (char)toupper(((int)*str));
         ++str;
     }
     return st;
